fix shine and fix crshing

+++ libft/libft.h
@@ -28,9 +28,9 @@ typedef struct		s_list
 
 typedef struct		s_vec
 {
-	float			x;
-	float			y;
-	float			z;
+	double			x;
+	double			y;
+	double			z;
 }					t_vec;
 
 int					ft_abs(int d);
@@ -104,12 +104,12 @@ void				ft_lstdelone(t_list **alst, void (*del)(void*, size_t));
 void				ft_lstadd(t_list **alst, t_list *new);
 void				ft_lstiter(t_list *lst, void (*f)(t_list *elem));
 t_list				*ft_lstmap(t_list *lst, t_list *(*f)(t_list *elem));
-t_vec				ft_vectorscale(t_vec *v, float n);
+t_vec				ft_vectorscale(t_vec *v, double n);
 t_vec				ft_vectorsub(t_vec *v1, t_vec *v2);
 t_vec				ft_vectorcross(t_vec *v1, t_vec *v2);
 void				ft_vectornorm(t_vec *v);
-float				ft_vectordot(t_vec *v1, t_vec *v2);
-float				ft_vectorlen(t_vec *v);
+double				ft_vectordot(t_vec *v1, t_vec *v2);
+double				ft_vectorlen(t_vec *v);
 t_vec				ft_vectoradd(t_vec *v1, t_vec *v2);
 int					ft_rgb_to_int(int red, int green, int blue);
